Changed implementation of EventMatchOperator and MinOccurEvent, such that EventMatchOperator counts event occurrences and does not ned to notify MinOccurEvent every time

+++ Communication/CommunicationBase/src/main/java/de/tum/in/i22/uc/cm/distribution/IDistributionManager.java
@@ -31,7 +31,7 @@ public interface IDistributionManager {
 
 	boolean wasTrueInBetween(AtomicOperator operator, long from, long to);
 
-	public long howOftenTrueInBetween(AtomicOperator operator, long from, long to);
+	public int howOftenTrueInBetween(AtomicOperator operator, long from, long to);
 
 	void unregisterPolicy(String policyName, IPLocation location);
